+++ changelog.md
@@ -3,7 +3,7 @@
 - Add an additional keybind to save and exit an editor level.
 - Add a keybind to open the Geode mods list, almost accessible anywhere.
 - Add a keybind to open this mod's settings, also almost accessible anywhere.
-- Add keybinds to (toggle between viewing/creating new) local editor levels and local level lists.
+- Add keybinds to (toggle between viewing/creating new/viewing uploaded) custom levels and level lists.
 - Remove forcing player death SFX settings (please use v1.1.0 of Adam729's Random Death Sounds mod instead).
 - Fix an issue where SFX in some levels would infinitely loop. Thank you, Aktimoose, for bringing this to my attention, and deepest apologies for rejecting your original claims. I hope this apology and this update is compensation enough for my harsh behavior at the time.
 - Added punctuation to changelogs.